D3ASIM-3553: Rename REST device client class
Use the new Asset name to identify devices
